refactor: update module layout to new style
This converts from <module_name>/mod.rs to simply <module_name>.rs.
Additionally changes the tests to be integration tests so they consume
the rust-ceramic-migrations-tests crate via its public interface. This
means we can remove the allow_dead code directive as the tests directly
consume the crate.
