+++ .dockerignore
@@ -1,4 +1,5 @@
 /target
+/test-binaries
 */target
 .github
 Dockerfile
