+++ .gitignore
@@ -3,3 +3,5 @@
 
 .DS_Store
 .idea
+
+/test-binaries
